backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Pydantic
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Redis
redis>=5.0.0

# Security
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
google-auth>=2.27.0
requests>=2.31.0

# Rate limiting
slowapi>=0.1.9

# File uploads
python-multipart>=0.0.6

# HTTP client
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
